requirements.txt
streamlit
streamlit-cropper
pillow
jinja2
